refactor: rename api security handlers
