+++ src/main/java/com/mogak/spring/security/ApiAuthenticationEntryPoint.java
@@ -1,7 +1,8 @@
-package com.mogak.spring.jwt;
+package com.mogak.spring.security;
 
 import com.mogak.spring.exception.ErrorResponse;
 import com.mogak.spring.global.ErrorCode;
+import com.mogak.spring.jwt.JwtAuthenticationException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.AuthenticationException;
@@ -12,17 +13,23 @@ import org.springframework.stereotype.Component;
 import java.io.IOException;
 
 @Component
-public final class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public final class ApiAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException)
             throws IOException {
-        ErrorCode errorCode = authException instanceof JwtAuthenticationException jwtAuthenticationException
-                ? jwtAuthenticationException.getErrorCode()
-                : authException instanceof BadCredentialsException
-                ? ErrorCode.WRONG_TOKEN
-                : ErrorCode.EMPTY_TOKEN;
+        ErrorCode errorCode = resolveErrorCode(authException);
         writeError(response, ErrorResponse.of(errorCode));
+    }
+
+    private ErrorCode resolveErrorCode(AuthenticationException authException) {
+        if (authException instanceof JwtAuthenticationException jwtAuthenticationException) {
+            return jwtAuthenticationException.getErrorCode();
+        }
+        if (authException instanceof BadCredentialsException) {
+            return ErrorCode.WRONG_TOKEN;
+        }
+        return ErrorCode.EMPTY_TOKEN;
     }
 
     private void writeError(HttpServletResponse response, ErrorResponse errorResponse) throws IOException {
